refactor : 안 쓰는 파일 및 import 삭제

+++ src/main/java/com/programmers/voucher/model/customer/Customer.java
@@ -1,17 +1,17 @@
 package com.programmers.voucher.model.customer;
 
 public class Customer {
-    private final int customerId;
+    private final long customerId;
     private final String customerName;
     private final String email;
 
-    public Customer(int customerId, String customerName, String email) {
+    public Customer(long customerId, String customerName, String email) {
         this.customerId = customerId;
         this.customerName = customerName;
         this.email = email;
     }
 
-    public int getCustomerId() {
+    public long getCustomerId() {
         return customerId;
     }
